myproject/requirements.txt
africastalking==1.2.9
alembic==1.13.1
aniso8601==9.0.1
anyio==4.8.0
apturl==0.5.2
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asgiref==3.8.1
asttokens==2.4.1
async-lru==2.0.4
attrs==24.3.0
Authlib==1.0.0
babel==2.16.0
bcrypt==3.2.0
beautifulsoup4==4.12.3
bleach==6.2.0
blinker==1.7.0
Brlapi==0.8.3
certifi==2020.6.20
cffi==1.17.1
chardet==4.0.0
charset-normalizer==3.4.1
click==8.1.7
colorama==0.4.4
comm==0.2.2
command-not-found==0.3
contourpy==1.3.1
cryptography==3.4.8
cupshelpers==1.0
cycler==0.12.1
dbus-python==1.2.18
debugpy==1.8.11
decorator==5.1.1
defer==1.0.6
defusedxml==0.7.1
distlib==0.3.8
distro==1.7.0
distro-info==1.1+ubuntu0.2
Django==5.1.1
django-cors-headers==4.4.0
django-phonenumber-field==8.0.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
docker==5.0.3
docker-compose==1.29.2
dockerpty==0.4.1
docopt==0.6.2
drf-yasg==1.21.8
duplicity==0.8.21
exceptiongroup==1.2.0
executing==2.0.1
Faker==24.1.0
fasteners==0.14.1
fastjsonschema==2.21.1
filelock==3.13.1
Flask==3.0.2
Flask-Bcrypt==1.0.1
Flask-Cors==4.0.0
Flask-JWT-Extended==4.6.0
Flask-Migrate==4.0.7
Flask-RESTful==0.3.10
Flask-SQLAlchemy==3.1.1
fonttools==4.55.3
fqdn==1.5.1
future==0.18.2
greenlet==3.0.3
gunicorn==22.0.0
h11==0.14.0
httpcore==1.0.7
httplib2==0.20.2
httpx==0.28.1
idna==3.3
importlib-metadata==4.6.4
inflection==0.5.1
iniconfig==2.0.0
ipdb==0.13.13
ipykernel==6.29.5
ipython==8.23.0
isoduration==20.11.0
itsdangerous==2.1.2
jedi==0.19.1
jeepney==0.7.1
Jinja2==3.1.3
jose==1.0.0
json5==0.10.0
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter-events==0.11.0
jupyter-lsp==2.2.5
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyter_server==2.15.0
jupyter_server_terminals==0.5.3
jupyterlab==4.3.4
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
keyring==23.5.0
kiwisolver==1.4.8
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lockfile==0.12.2
louis==3.20.0
macaroonbakery==1.3.1
Mako==1.1.3
MarkupSafe==2.1.5
marshmallow==3.21.1
marshmallow-sqlalchemy==1.0.0
matplotlib==3.10.0
matplotlib-inline==0.1.7
mistune==3.1.0
monotonic==1.6
more-itertools==8.10.0
natsort==8.0.2
nbclient==0.10.2
nbconvert==7.16.5
nbformat==5.10.4
nest-asyncio==1.6.0
netifaces==0.11.0
notebook==7.3.2
notebook_shim==0.2.4
numpy==2.2.1
oauthlib==3.2.0
olefile==0.46
overrides==7.7.0
packaging==23.2
pandas==2.2.3
pandocfilters==1.5.1
paramiko==2.9.3
parso==0.8.4
pexpect==4.8.0
phonenumbers==8.13.53
Pillow==9.0.1
pipenv==2023.12.1
platformdirs==4.2.0
plotly==5.24.1
pluggy==1.4.0
prometheus_client==0.21.1
prompt-toolkit==3.0.43
protobuf==3.12.4
psutil==6.1.1
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pure-eval==0.2.2
pycairo==1.20.1
pycparser==2.22
pycups==2.0.1
Pygments==2.17.2
PyGObject==3.42.1
PyJWT==2.10.1
pymacaroons==0.13.0
PyNaCl==1.5.0
pyparsing==2.4.7
pyRFC3339==1.1
pyrsistent==0.18.1
pytest==8.0.2
python-apt==2.4.0+ubuntu4
python-dateutil==2.9.0.post0
python-debian==0.1.43+ubuntu1.1
python-decouple==3.8
python-dotenv==1.0.1
python-json-logger==3.2.1
python3-openid==3.2.0
pytz==2022.1
pyxdg==0.27
PyYAML==5.4.1
pyzmq==26.2.0
referencing==0.35.1
reportlab==3.6.8
requests==2.32.3
requests-oauthlib==2.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.22.3
schema==0.7.7
seaborn==0.13.2
SecretStorage==3.3.1
Send2Trash==1.8.3
six==1.16.0
sniffio==1.3.1
social-auth-app-django==5.4.2
social-auth-core==4.5.4
soupsieve==2.6
SQLAlchemy==2.0.29
SQLAlchemy-serializer==1.4.12
sqlparse==0.5.1
stack-data==0.6.3
systemd-python==234
tenacity==9.0.0
terminado==0.18.1
texttable==1.6.4
tinycss2==1.4.0
tomli==2.0.1
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.2
types-python-dateutil==2.9.0.20241206
typing_extensions==4.10.0
tzdata==2024.2
ubuntu-drivers-common==0.0.0
ubuntu-pro-client==8001
ufw==0.36.1
unattended-upgrades==0.1
uri-template==1.3.0
uritemplate==4.1.1
urllib3==1.26.5
usb-creator==0.3.7
vboxapi==1.0
virtualenv==20.25.1
virtualenv-clone==0.3.0
wadllib==1.3.6
wcwidth==0.2.13
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
Werkzeug==3.0.2
wxPython==4.0.7
xdg==5
xkit==0.0.0
zipp==1.0.0
